+++ src/main/java/com/swp391/OnlineEnglishLearningSystem/service/SliderService.java
@@ -10,11 +10,11 @@ import java.util.List;
 public interface SliderService {
     Page<Slider> getSliders(String keyword, String status, int page, int size);
     List<Slider> getActiveSliders();
-    Slider getSliderById(Integer id);
+    Slider getSliderById(Long id);
     Slider createSlider(SliderCreateUpdateDto dto);
-    Slider updateSlider(Integer id, SliderDTO dto);
-    Slider updateSliderWithFile(Integer id, SliderCreateUpdateDto dto);
-    Slider toggleStatus(Integer id, String status);
-    void deleteSlider(Integer id);
-    void incrementViewCount(Integer id);
+    Slider updateSlider(Long id, SliderDTO dto);
+    Slider updateSliderWithFile(Long id, SliderCreateUpdateDto dto);
+    Slider toggleStatus(Long id, String status);
+    void deleteSlider(Long id);
+    void incrementViewCount(Long id);
 }
